requirements.txt
# K-POP Schedule Auto-Feed 依存関係
# 現在動作中の最小構成 + 将来実装予定の機能用

# === 現在使用中（最小構成） ===
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# === 将来実装予定の機能用 ===
# Google Cloud関連
google-cloud-firestore>=2.15.0
google-cloud-aiplatform>=1.30.0
vertexai>=1.30.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.1.0
functions-framework>=3.5.0

# スクレイピング・データ取得
snscrape>=0.7.0
serpapi>=0.1.5
requests>=2.30.0
beautifulsoup4>=4.12.0

# 日本語処理
jaconv>=0.3.0
mojimoji>=0.0.12

# カレンダー連携
icalendar>=5.0.0

# データ処理
pandas>=2.0.0
numpy>=1.24.0

# Cloud Run / Functions対応
mangum>=0.17.0

# 開発・テスト用
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
responses>=0.25.0
httpx>=0.25.0
google-generativeai>=0.7.0
google-cloud-firestore>=2.15.0